+++ osu.Desktop.VisualTests/Tests/TestCasePlayer.cs
@@ -15,6 +15,7 @@ using osu.Game.Modes.Objects;
 using osu.Game.Modes.Osu.Objects;
 using osu.Game.Screens.Play;
 using OpenTK.Graphics;
+using osu.Desktop.VisualTests.Beatmaps;
 
 namespace osu.Desktop.VisualTests.Tests
 {
@@ -96,17 +97,6 @@ namespace osu.Desktop.VisualTests.Tests
             {
                 Beatmap = beatmap
             };
-        }
-
-        private class TestWorkingBeatmap : WorkingBeatmap
-        {
-            public TestWorkingBeatmap(Beatmap beatmap)
-                : base(beatmap.BeatmapInfo, beatmap.BeatmapInfo.BeatmapSet)
-            {
-                Beatmap = beatmap;
-            }
-
-            protected override ArchiveReader GetReader() => null;
         }
     }
 }
